Allow creating clients in Rails that do not automatically inject themselves.
